updated get-sessions endpoint
